Add pythonpath configuration (#37)
Name the pythonpath configuration after the `pythonpath` cli flag `--pythonpath`. Always configure the pythonpath if there is a pyproject.toml, a breaking change.

Rename the `settings_module` configuration to `settings` to match the cli flag `-- settings`. The old name is preserved for backward compatibility.

Deprecate using `setup.cfg`. It is kept for backward compatibility, but does not have feature parity with `pyproject.toml`.